requirements.txt
amqp==2.2.1
appnope==0.1.0
asn1crypto==0.22.0
autopep8==1.3.2
bcrypt==3.1.3
billiard==3.5.0.3
bleach==2.0.0
celery==4.0.2
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
click==6.7
cryptography==2.0
decorator==4.1.1
entrypoints==0.2.3
Fabric3==1.13.1.post1
Flask==0.12.2
Flask-SQLAlchemy==2.2
html5lib==0.999999999
idna==2.5
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.1.0
jupyter-core==4.3.0
kombu==4.0.2
lxml==3.8.0
MarkupSafe==1.0
mistune==0.7.4
nbconvert==5.2.1
nbformat==4.3.0
notebook==5.0.0
pandocfilters==1.4.1
paramiko==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.2
pyasn1==0.2.3
pycodestyle==2.3.1
pycparser==2.18
Pygments==2.2.0
PyNaCl==1.1.2
python-dateutil==2.6.1
pytz==2017.2
pyzmq==16.0.2
qtconsole==4.3.0
requests==2.18.1
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.11
terminado==0.6
testpath==0.3.1
tornado==4.5.1
traitlets==4.3.2
urllib3==1.21.1
vine==1.1.4
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
widgetsnbextension==2.0.0
